+++ src/google/protobuf/reflection_ops.cc
@@ -63,7 +63,7 @@ void ReflectionOps::Merge(const Message& from, Message* to) {
 
   vector<const FieldDescriptor*> fields;
   from_reflection->ListFields(from, &fields);
-  for (int i = 0; i < fields.size(); i++) {
+  for (size_t i = 0; i < fields.size(); i++) {
     const FieldDescriptor* field = fields[i];
 
     if (field->is_repeated()) {
@@ -129,7 +129,7 @@ void ReflectionOps::Clear(Message* message) {
 
   vector<const FieldDescriptor*> fields;
   reflection->ListFields(*message, &fields);
-  for (int i = 0; i < fields.size(); i++) {
+  for (size_t i = 0; i < fields.size(); i++) {
     reflection->ClearField(message, fields[i]);
   }
 
@@ -152,7 +152,7 @@ bool ReflectionOps::IsInitialized(const Message& message) {
   // Check that sub-messages are initialized.
   vector<const FieldDescriptor*> fields;
   reflection->ListFields(message, &fields);
-  for (int i = 0; i < fields.size(); i++) {
+  for (size_t i = 0; i < fields.size(); i++) {
     const FieldDescriptor* field = fields[i];
     if (field->cpp_type() == FieldDescriptor::CPPTYPE_MESSAGE) {
 
@@ -183,7 +183,7 @@ void ReflectionOps::DiscardUnknownFields(Message* message) {
 
   vector<const FieldDescriptor*> fields;
   reflection->ListFields(*message, &fields);
-  for (int i = 0; i < fields.size(); i++) {
+  for (size_t i = 0; i < fields.size(); i++) {
     const FieldDescriptor* field = fields[i];
     if (field->cpp_type() == FieldDescriptor::CPPTYPE_MESSAGE) {
       if (field->is_repeated()) {
@@ -238,7 +238,7 @@ void ReflectionOps::FindInitializationErrors(
   // Check sub-messages.
   vector<const FieldDescriptor*> fields;
   reflection->ListFields(message, &fields);
-  for (int i = 0; i < fields.size(); i++) {
+  for (size_t i = 0; i < fields.size(); i++) {
     const FieldDescriptor* field = fields[i];
     if (field->cpp_type() == FieldDescriptor::CPPTYPE_MESSAGE) {
