+++ src/google/protobuf/wire_format.cc
@@ -710,7 +710,7 @@ void WireFormat::SerializeWithCachedSizes(
 
   vector<const FieldDescriptor*> fields;
   message_reflection->ListFields(message, &fields);
-  for (int i = 0; i < fields.size(); i++) {
+  for (size_t i = 0; i < fields.size(); i++) {
     SerializeFieldWithCachedSizes(fields[i], message, output);
   }
 
@@ -880,7 +880,7 @@ int WireFormat::ByteSize(const Message& message) {
 
   vector<const FieldDescriptor*> fields;
   message_reflection->ListFields(message, &fields);
-  for (int i = 0; i < fields.size(); i++) {
+  for (size_t i = 0; i < fields.size(); i++) {
     our_size += FieldByteSize(fields[i], message);
   }
